+++ src/component/bar/Bar.module.scss
