+++ src/component/bar/barContent/BarContent.module.scss
