+++ src/component/bar/barContent/Player.module.scss
