+++ src/component/bar/barContent/playerBlock/player/Player.jsx
