+++ src/component/bar/barContent/playerProgress/PlayerProgress.jsx
